api/requirements.txt
Django>=5.2.6,<5.3
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=24.2,<25.0
psycopg[binary]>=3.2,<3.3
gunicorn>=23,<24
whitenoise>=6.7,<7.0
python-dotenv>=1.0

drf-spectacular
drf-spectacular-sidecar
requests>=2.32
